src/requirements.txt
Django
South==0.8.4
Unipath==1.0
argparse==1.2.1
dj-database-url==0.2.2
dj-static==0.0.5
django-decouple==2.1
djangorestframework==2.3.13
gunicorn==18.0
psycopg2==2.5.2
pystache==0.5.3
static==1.0.2
wsgiref==0.1.2
pytz==2014.2

#local package
django-debug-toolbar==1.1


#fork package
#django-grappelli==2.5.2
#django-tenant-schemas==1.4.5
#django-plans
	#|_suds==0.4 --> is a lightweight SOAP python client that provides a service proxy for Web services.
	#|_django-ordered-model==0.3.0
	#|_django-countries==2.1.2
	#|_suds
	#|_six
	#|_vatnumber

django-countries
django-ordered-model
suds
six
vatnumber